crawler/ctrip/requirements.txt
selenium
requests
pymongo
Pillow
